requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
pandas>=2.0
openpyxl>=3.1
Pillow>=10.0.0
